% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PRISTDataPackage.R
\docType{package}
\name{MASTDataPackage}
\alias{MASTDataPackage}
\alias{MASTDataPackage-package}
\title{Data package for Single Cell RNA sequence paper}
\description{
MASTDataPackage for MAST paper
}
\details{
This package contains data used for MAST paper.
}
\author{
Jingyuan Deng, Greg Finak, Andrew McDavid, Masaano Yajima
}

