% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PRISTDataPackage.R
\docType{package}
\name{PRISTDataPackage}
\alias{PRISTDataPackage}
\alias{PRISTDataPackage-package}
\title{Data package for Single Cell RNA sequence paper}
\description{
PRISTDataPackage for PRIST paper
}
\details{
This package contains data used for PRIST paper.
}
\author{
Jingyuan Deng, Greg Finak, Andrew McDavid, Masaano Yajima
}

