% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MASTDataPackage.R, R/PRISTDataPackage.R
\docType{data}
\name{sca_alex}
\alias{sca_alex}
\title{Create Alex data}
\format{This is a SingleCellAssay object
\describe{
\itemize{
\item{cData}{Cell level feature annotation.}
\item{wellKey} unique identifier
\item{Stim} Stimulation
\item{Time} Time
\item{BioRep} Boolean for biological replicates
\item{TechRep} Boolean for technical replicates
\item{OtherCondition} Boolean for other conditions
\item{ngeneson} percentage of genes that are turned on
\item{cngeneson} centered ngeneson
	}
\item{fData}{Gene level feature annotation.}
\itemize{
\item{primerid} gene symbols
\item{GENES} gene symbols
	}
\item{exprs}{Expression matrix}
}}
\source{
The data was from Matin

The data was from Matin
}
\description{
Processed single cell RNA seqdata for our single cell RNA seq paper from Alex study.

Processed single cell RNA seqdata for our single cell RNA seq paper from Alex study.
}
\details{
This data set contains the processed data for our fake
assay, with 100 replicated measurements for each of four subjects.

This data set contains the processed data for our fake
assay, with 100 replicated measurements for each of four subjects.
}
\note{
We used UCSC hg19 as the reference genome and used RSEM (bowtie) to align and calculate the count and the TPM matrix.

We used UCSC hg19 as the reference genome and used RSEM (bowtie) to align and calculate the count and the TPM matrix.
}
\author{
Jingyuan Deng, Masanao Yajima, Andrew McDavid, Greg Finak

Jingyuan Deng, Masanao Yajima, Andrew McDavid, Greg Finak
}

