% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PRISTDataPackage.R
\docType{data}
\name{sca_mait}
\alias{sca_mait}
\title{Create MAIT data}
\format{This is a SingleCellAssay object
\describe{
\item{cData}{Cell level feature annotation.}
\itemize{
\item{wellKey} unique identifier
\item{condition} stimulated or unstimulated
\item{ncells} number of cells in the well
\item{ngeneson} percentage of genes that are turned on
\item{cngeneson} centered ngeneson
\item{TRAV1} TRAV1
\item{TRBV6} TRBV6
\item{TRBV4} TRBV4
\item{TRBV20} TRBV20
\item{alpha} alpha chain
\item{beta}  beta chain
\item{ac}    alpha chain + stimulation
\item{bc}    beta chain + stimulation
	}
\item{fData}{Gene level feature annotation.}s
\itemize{
\item{primerid} entrez id
\item{entrez} entrez id
\item{symbolid} gene symbols
	}
\item{exprs}{Expression matrix}
}}
\source{
The data was from Matin
}
\description{
Processed single cell RNA seqdata for our single cell RNA seq paper from our MAIT study.
}
\details{
This data set contains the processed data for our fake
assay, with 100 replicated measurements for each of four subjects.
}
\note{
We used UCSC hg19 as the reference genome and used RSEM (bowtie) to align and calculate the count and the TPM matrix.
}
\author{
Jingyuan Deng, Masanao Yajima, Andrew McDavid, Greg Finak
}

